feat: OrderService
- 주문하기 및 주문내역 조회를 위한 서비스 로직 구현

+++ kakao-shopping/src/main/java/com/kakao/shopping/dto/order/OrderDTO.java
@@ -3,8 +3,8 @@ package com.kakao.shopping.dto.order;
 import java.util.List;
 
 public record OrderDTO(
-        int id,
+        Long id,
         List<OrderProductDTO> products,
-        int totalPrice
+        Long totalPrice
 ) {
 }
