+++ kakao-shopping/src/main/java/com/kakao/shopping/dto/order/OrderItemDTO.java
@@ -2,7 +2,7 @@ package com.kakao.shopping.dto.order;
 
 public record OrderItemDTO(
         String name,
-        int quantity,
-        int price
+        Long quantity,
+        Long price
 ) {
 }
