+++ kakao-shopping/src/main/java/com/kakao/shopping/dto/order/OrderUpdateRequest.java
@@ -1,8 +1,8 @@
 package com.kakao.shopping.dto.order;
 
 public record OrderUpdateRequest(
-        int orderId,
-        int optionId,
-        int quantity
+        Long orderId,
+        Long optionId,
+        Long quantity
 ) {
 }
